+++ mbtileReader.js
@@ -42,12 +42,12 @@ function readMetadata(file) {
 
 function listFiles(file, filter) {
 	log.info("Open " + file);
-	if (filter.length === 2) filter[1] = Math.pow(2, filter[0]) - 1 - filter[1];
+	if (filter.length === 3) filter[2] = Math.pow(2, filter[0]) - 1 - filter[2];
 	return new Promise((resolve, reject) => {
 		const sql = {
 			0: "SELECT DISTINCT zoom_level FROM tiles",
-			1: "SELECT (2 << zoom_level - 1) - 1 - tile_row FROM tiles WHERE zoom_level=?",
-			2: "SELECT tile_column, length(tile_data) AS size FROM tiles WHERE zoom_level=? AND tile_row=?"
+			1: "SELECT DISTINCT tile_column FROM tiles WHERE zoom_level=?",
+			2: "SELECT (2 << zoom_level - 1) - 1 - tile_row, length(tile_data) AS size FROM tiles WHERE zoom_level=? AND tile_column=?"
 		};
 
 		const db = new sqlite3.Database(file, sqlite3.OPEN_READONLY, err => {
