+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "tiny-tileserver",
   "description": "Tiny raster and vector tile server.",
-  "version": "1.1.0",
+  "version": "1.1.1",
   "main": "index.js",
   "devDependencies": {
     "eslint": "^5.6.0"
